Restructuring  of new chemical generator
- Moved database formula sampling  into own class
- Moved Formula and DummyFormula into Common to avoid infinite loop importing
- Created DummyFormula to allow a single mz value to be interpreted as a Formula thus making sampling simpler -- everything can be considered a Formula sampler
- Added a check in the chemical sampler that makes either a KnownChemical or UnknownChemical depending on whether the formula is a Formula or a DummyFormula
